Lots of false positives from decompilation differences, especially
in branching.  Some of them i accomodated by hand, others i didn't
touch.  Cleaned up diffs only of code that was simple enough to
confidently edit and also was likely to be touched by the mod.

+++ BMSymbol.cs
@@ -77,7 +77,7 @@ public class BMSymbol
 				}
 				else
 				{
-					mUV = new Rect((float)mSprite.x, (float)mSprite.y, (float)mSprite.width, (float)mSprite.height);
+					mUV = new Rect(mSprite.x, mSprite.y, mSprite.width, mSprite.height);
 					mUV = NGUIMath.ConvertToTexCoords(mUV, texture.width, texture.height);
 					mOffsetX = mSprite.paddingLeft;
 					mOffsetY = mSprite.paddingTop;
